requirements.txt
# GUI Framework
PyQt6>=6.4.0

# Computer Vision and Image Processing
opencv-python>=4.7.0
mtcnn>=0.1.1

# Deep Learning Frameworks
tensorflow==2.16.1
tf-keras
torch>=2.0.0
torchvision>=0.15.0
numba==0.59.1

# Scientific Computing
numpy>=1.24.0

# Video Processing
moviepy>=1.0.3

# Audio and Speech Recognition
openai-whisper>=20230314

# Natural Language Processing
transformers>=4.28.0
keybert>=0.7.0
sentence-transformers>=2.2.0

# HTTP Requests
requests>=2.28.0

# Environment Variables
python-dotenv>=1.0.0

# Additional Dependencies
Pillow>=9.5.0
scipy>=1.10.0
scikit-learn>=1.2.0